+++ src/main/java/com/ttasjwi/cafe/controller/HomeController.java
@@ -1,7 +1,7 @@
 package com.ttasjwi.cafe.controller;
 
 import com.ttasjwi.cafe.domain.Article;
-import com.ttasjwi.cafe.repository.ArticleRepository;
+import com.ttasjwi.cafe.service.ArticleService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -11,15 +11,15 @@ import java.util.List;
 @Controller
 public class HomeController {
 
-    private final ArticleRepository articleRepository;
+    private final ArticleService articleService;
 
-    public HomeController(ArticleRepository articleRepository) {
-        this.articleRepository = articleRepository;
+    public HomeController(ArticleService articleService) {
+        this.articleService = articleService;
     }
 
     @GetMapping
     public String home(Model model) {
-        List<Article> articleList = articleRepository.findAll();
+        List<Article> articleList = articleService.findArticles();
         model.addAttribute("articles", articleList);
         return "home";
     }